requirements.txt
aiogram==3.0.0
SQLAlchemy==2.0.20
psycopg2==2.9.7
python-dotenv==1.0.0